сделал пару тестов для HTTP запросов
