+++ lesson6/blog-app-mongo/tests/blog_load_test.go
@@ -1,0 +1,34 @@
+package test
+
+import (
+	"github.com/astaxie/beego"
+	. "github.com/smartystreets/goconvey/convey"
+	_ "go_web/lesson6/blog-app-mongo/routers"
+	"net/http"
+	"net/http/httptest"
+	"path/filepath"
+	"runtime"
+	"testing"
+)
+
+func init() {
+	_, file, _, _ := runtime.Caller(0)
+	apppath, _ := filepath.Abs(filepath.Dir(filepath.Join(file, ".."+string(filepath.Separator))))
+	beego.TestBeegoInit(apppath)
+}
+
+func TestBlogLoadOK(t *testing.T) {
+	r, _ := http.NewRequest("GET", "/", nil)
+	w := httptest.NewRecorder()
+
+	beego.BeeApp.Handlers.ServeHTTP(w, r)
+
+	Convey("Subject: Test Station Endpoint\n", t, func() {
+		Convey("Status Code Should Be 200", func() {
+			So(w.Code, ShouldEqual, 200)
+		})
+		Convey("The Result Should Not Be Empty", func() {
+			So(w.Body.Len(), ShouldBeGreaterThan, 0)
+		})
+	})
+}
